Validando todos los inputs

+++ src/app/modules/inputFilter.js
@@ -14,7 +14,7 @@ export function getInput(data) {
   /** Dato de input obtenido por el parámetro 'data' */
   const dataInput = document.querySelector(data);
   /** Valor de 'dataInput' convertido de tipo de dato 'String' a 'Number' */
-  const filteredInput = Number.parseInt(dataInput.value);
+  const filteredInput = Number.parseFloat(dataInput.value);
 
   return filteredInput;
 }
